backend/benefit/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
-e ../shared
    # via -r requirements.in
asgiref==3.3.4
    # via django
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
cryptography==3.4.7
    # via django-cryptography
django-appconf==1.0.4
    # via django-cryptography
django-cors-headers==3.7.0
    # via -r requirements.in
django-cryptography==1.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-localflavor==3.1
    # via -r requirements.in
django-simple-history==3.0.0
    # via -r requirements.in
django==3.2
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-cors-headers
    #   django-localflavor
    #   djangorestframework
    #   yjdh-backend-shared
djangorestframework==3.12.4
    # via -r requirements.in
idna==2.10
    # via requests
psycopg2==2.8.6
    # via -r requirements.in
pycparser==2.20
    # via cffi
python-stdnum==1.16
    # via django-localflavor
pytz==2021.1
    # via django
requests==2.25.1
    # via -r requirements.in
sentry-sdk==1.0.0
    # via -r requirements.in
sqlparse==0.4.1
    # via django
urllib3==1.26.4
    # via
    #   requests
    #   sentry-sdk
